requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.3.0